fixed conversion, checking over and underflow

+++ src/PixelHistogram.test.cpp
@@ -13,7 +13,7 @@ using aare::NDView;
 
 TEST_CASE("Fill one pixel of a 5x10 histogram"){
     PixelHistogram hist(5, 10, 20, 0.0, 10.0);
-    NDArray<double, 2> image({5, 10}, -1.0); //Need to fill with -1 to not generate counts
+    NDArray<float, 2> image({5, 10}, -1.0); //Need to fill with -1 to not generate counts
     
     image(2, 3) = 5.7; // This should go into bin 11 (since bins are [0-0.5), [0.5-1.0), ..., [9.5-10.0))
     
@@ -41,7 +41,7 @@ TEST_CASE("Fill one pixel of a 5x10 histogram"){
 
 TEST_CASE("Fill pixels with uneven partial histogram row slices"){
     PixelHistogram hist(5, 4, 10, 0.0, 10.0, 3);
-    NDArray<double, 2> image({5, 4}, -1.0);
+    NDArray<float, 2> image({5, 4}, -1.0);
 
     image(0, 0) = 0.2;
     image(1, 1) = 1.2;
